Bug fix: support redirects with location that has unicode characters.
If the redirect location has a unicode character like é, then the subsequent request will not be url encoded, and will therefor fail with a not found error.

This usecase happens in phonegap build, for CLI version 7.1.0 and presumably higher, when the name of the ios application has a unicode character. Then the download link will be a redirect to a location on S3 that thas the problematic character.

Using the more modern node URL interface to parse the location on redirect handles the encoding.